🧪 [TEST] Missing test file for src/godot/headless.ts
🎯 What:
- Consolidated redundant test files into a single, comprehensive \`tests/godot/headless.test.ts\`.
- Added new test cases to cover all branches in \`src/godot/headless.ts\`, including error handling edge cases.
- Fixed Biome linting issues by replacing \`@ts-ignore\` with \`@ts-expect-error\`.

💡 Why:
- The previous test setup was fragmented and incomplete.
- Branches in \`headless.ts\` were untested (achieved 100% branch coverage).
- Consolidation improves maintainability and ensures a single source of truth for headless execution tests.

✅ Verification:
- Ran \`npx vitest run tests/godot/headless.test.ts --coverage\` and confirmed 100% coverage.
- Ran the full project test suite (\`bun run test\`): 650/650 passed.
- Ran \`bun run check\`: Passed (formatting and types).

✨ Result:
- Achievement of 100% code coverage for the Godot headless execution module.
- Cleaner and more robust test suite.
